refactor(swarm): split supervisor worker and probe helpers
Keep worker lifecycle and verification logic in dedicated modules so supervisor.py stays patchable but easier to navigate and maintain.
